+++ plugins/multi-chain-web-authn/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @zerodev/multi-chain-web-auth
 
+## 5.5.0
+
+### Minor Changes
+
+- feat: security patch
+
 ## 5.4.10
 
 ### Patch Changes