+++ plugins/permission/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @zerodev/permissions
 
+## 5.6.2
+
+### Patch Changes
+
+- feat: added migrationSessionKeyAccount and isPermissionMigrated utils
+
+## 5.6.1
+
+### Patch Changes
+
+- feat: added optional permissionId param in toPermissionValidator and added it as serialization param
+
+## 5.6.0
+
+### Minor Changes
+
+- feat: security patch
+
 ## 5.5.14
 
 ### Patch Changes